remove reference to broken jquery asset
We're already loading jquery in _includes/scripts.html via the CDN
(https://code.jquery.com/jquery-3.3.1.min.js), so we don't need this
additional reference, which was unsuccessfully trying to load a local
version of jquery.

As a result of this change, there should no longer be any errors in the
console or network panel.